added post route simulation testbench for infer_dual_port_rom

+++ Testcases/infer_dual_port_rom/bit_sim.sh
@@ -4,6 +4,7 @@
 main_path=$PWD
 
 design_name=${PWD##*/}
+simulator_name="iverilog" #vcs,iverilog
 
 device=GEMINI_COMPACT_10x8
 
@@ -75,9 +76,9 @@ cd $main_path
 # python3 ../../scripts/gen_openfpga_script.py $design_name $vpr_file $openfpga_file $fixed_sim_openfpga_file $repack_design_constraint_file $bitstream_annotation_file $default
 
 design_path=`find . -type f -iname "$design_name.v"`
-tool_name="vcs"
+tool_name="iverilog"
 
-sed -i -e "s|MEM_FILE_PATH|$PWD/rtl|g" rtl/rom.v
+# sed -i -e "s|MEM_FILE_PATH|$PWD/rtl|g" rtl/rom.v
 
 command -v raptor >/dev/null 2>&1 && raptor_path=$(which raptor) || { echo >&2 echo "First you need to source Raptor"; end_time exit; }
 lib_fix_path="${raptor_path:(-11)}"
@@ -163,19 +164,51 @@ TDP18K_FIFO=`find $library -wholename "*/genesis3/TDP18K_FIFO.v"`
 ufifo_ctl=`find $library -wholename "*/genesis3/ufifo_ctl.v"`
 sram1024x18=`find $library -wholename "*/genesis3/sram1024x18.v"`
 primitive=`find $library -wholename "*/genesis3/primitives.v"`
+DFFRE=`find $library -wholename "*/FPGA_PRIMITIVES_MODELS/sim_models/verilog/DFFRE.v"`
+TDP_RAM18KX2=`find $library -wholename "*/FPGA_PRIMITIVES_MODELS/sim_models/verilog/TDP_RAM18KX2.v"`
+rs_tdp36k=`find $library -wholename "*/FPGA_PRIMITIVES_MODELS/sim_models_internal/primitives_mapping/BRAM/rs_tdp36k_post_pnr_mapping.v"`
 
-[ ! -d $design_name\_$tool_name\_post_route_files ] && mkdir $design_name\_$tool_name\_post_route_files
-[ -d $design_name\_$tool_name\_post_route_files ] && cd $design_name\_$tool_name\_post_route_files
-start_post_route=`date +%s`
-timeout 4m vcs -sverilog -timescale=1ns/1ps $cell_path $bram_sim $lut_map $TDP18K_FIFO $ufifo_ctl $sram1024x18 $dsp_sim $primitive ../../rtl/$design_name.v $post_route_netlist_path $route_tb_path +incdir+$directory_path -y $directory_path +libext+.v +define+VCS_MODE=1 -full64 -debug_all -lca -kdb | tee post_route_sim.log
-./simv | tee -a post_route_sim.log
-end_post_route=`date +%s`
-runtime_post_route=$((end_post_route-start_post_route))
-echo -e "\nTotal RunTime: $runtime_post_route sec">>post_route_sim.log
+python3 ../../../scripts/post_route_script.py $design_name 2
 
+if [[ $simulator_name == "vcs" ]]
+then
+    [ ! -d $design_name\_$tool_name\_post_route_files ] && mkdir $design_name\_$tool_name\_post_route_files
+    [ -d $design_name\_$tool_name\_post_route_files ] && cd $design_name\_$tool_name\_post_route_files
+    start_post_route=`date +%s`
+    timeout 4m vcs -sverilog -timescale=1ns/1ps $cell_path $bram_sim $lut_map $TDP18K_FIFO $ufifo_ctl $sram1024x18 $dsp_sim $primitive ../../rtl/$design_name.v $post_route_netlist_path $route_tb_path +incdir+$directory_path -y $directory_path +libext+.v +define+VCS_MODE=1 -full64 -debug_all -lca -kdb | tee post_route_sim.log
+    ./simv | tee -a post_route_sim.log
+    end_post_route=`date +%s`
+    runtime_post_route=$((end_post_route-start_post_route))
+    echo -e "\nTotal RunTime: $runtime_post_route sec">>post_route_sim.log
+fi
+
+if [[ $simulator_name == "iverilog" ]]
+then
+    [ ! -d $design_name\_$simulator_name\_post_route_files ] && mkdir $design_name\_$simulator_name\_post_route_files
+    [ -d $design_name\_$simulator_name\_post_route_files ] && cd $design_name\_$simulator_name\_post_route_files
+    cp ../../rtl/memory_test.mem .
+    start_post_route=`date +%s`
+    iverilog -g2012 -DIVERILOG=1 -o $design_name $TDP_RAM18KX2 $rs_tdp36k $primitive ../../rtl/$design_name.v $post_route_netlist_path $route_tb_path -y $main_path/rtl && vvp ./$design_name | tee post_route_sim.log
+    end_post_route=`date +%s`
+    runtime_post_route=$((end_post_route-start_post_route))
+    echo -e "\nTotal RunTime: $runtime_post_route sec">>post_route_sim.log
+fi
+
+if [[ $simulator_name == "verilator" ]]
+then
+    [ ! -d $design_name\_$simulator_name\_post_route_files ] && mkdir $design_name\_$simulator_name\_post_route_files
+    [ -d $design_name\_$simulator_name\_post_route_files ] && cd $design_name\_$simulator_name\_post_route_files
+    start_post_route=`date +%s`
+    # verilator -Wno-fatal -Wno-BLKANDNBLK -main -build -exe $route_tb_path ../../sim/post_route_tb/tb_sim_route_rom.cpp --timing --timescale 1ps/1ps --top-module rom_post_route_tb --trace -v $bram_sim -v $TDP18K_FIFO -v $ufifo_ctl -v $sram1024x18 $main_path/rtl/$design_name.v -v $primitive $post_route_netlist_path +libext+.v+.sv && make -j -C obj_dir -f Vrom_post_route_tb.mk 
+    # ./obj_dir/Vrom_post_route_tb >> post_route_sim.log
+    end_post_route=`date +%s`
+    runtime_post_route=$((end_post_route-start_post_route))
+    echo -e "\nTotal RunTime: $runtime_post_route sec">>post_route_sim.log
+fi
 while read line; do
         if [[ $line == *"All Comparison Matched"* ]]
         then
+            vcd2fst tb.vcd tb.fst --compress
             rm -fr tb.vcd
         fi
         if [[ $line == *"Error: Simulation Failed"* ]]
@@ -200,6 +233,8 @@ cd $design_name/$design_name\_golden/$design_name\_$tool_name\_bitstream_sim_fil
 python3 ../../../../scripts/force.py $design_name
 
 start_bitstream=`date +%s`
+# timeout 20m vcs -sverilog $bitstream_tb_path -full64 -debug_all -lca -kdb | tee bitstream_sim.log
+# ./simv | tee -a bitstream_sim.log
 end_bitstream=`date +%s`
 runtime_bitstream=$((end_bitstream-start_bitstream))
 echo -e "\nTotal RunTime: $runtime_bitstream sec">>bitstream_sim.log
